Moved CallFlowControl and CDRServer into a shared project.
