--- CDRServer/pubspec.yaml
@@ -1,16 +1,0 @@
-name: MessageServer
-description: A simple REST-based message-store server
-dependencies:
-  args: any
-  http: any
-  intl: any
-  mailer: any
-  postgresql: '>= 0.3.0'
-  openreception_framework:
-    version: ">= 0.0.13+0"
-    git:
-      ref: master
-      url: https://github.com/BitStackers/OpenReceptionFramework.git
-  path: any
-  route: any
-  xml: 1.0.0
