--- CallFlowControl/pubspec.yaml
@@ -1,22 +1,0 @@
-name: CallFlowControl
-version: 0.1.1
-author: Kim Rostgaard Christensen
-description: Call-Flow-Control server.
-dependencies:
-  args: any
-  intl: any
-  esl:
-    git:
-      ref: master
-      url: https://github.com/rostgaard/libesl-dart.git
-
-  openreception_framework:
-    version: ">= 0.0.15+7"
-    git:
-      ref: master
-      url: https://github.com/BitStackers/OpenReceptionFramework.git
-  path: any
-  http: any
-  route: any
-  logging: any
-
